requirements.txt
Pillow==8.4.0
numpy==1.21.5
tqdm==4.62.3
json
